refactor: put reading components in lib/component to be reusable
