+++ src/lib/components/reading/reading_quiz_finished_image.png
